Changes in reaction to review
The only task of BuildConceptTimeValuesStep seems to be to call ConceptTimeValuesTable, so I merged that
functionality into one class.

The optionality of building the table is now implemented by not creating the step that does it.
