+++ src/groovy/jobs/steps/BuildConceptTimeValuesStep.groovy
@@ -1,13 +1,19 @@
 package jobs.steps
 
 import au.com.bytecode.opencsv.CSVWriter
+import org.springframework.beans.factory.annotation.Autowired
+import org.transmartproject.core.ontology.ConceptsResource
+import org.transmartproject.core.ontology.OntologyTerm
 
 /**
  * Created by carlos on 1/27/14.
  */
 class BuildConceptTimeValuesStep implements Step {
 
-    Map<String,Map> table
+    @Autowired
+    ConceptsResource conceptsResource
+
+    List<String> timeValuesConcepts
 
     String[] header
 
@@ -23,6 +29,8 @@ class BuildConceptTimeValuesStep implements Step {
 
         //makes sure the file is not there
         outputFile.delete()
+
+        def table = computeMap(timeValuesConcepts)
 
         if (table != null) {
             writeToFile(table)
@@ -41,5 +49,36 @@ class BuildConceptTimeValuesStep implements Step {
             }
         }
     }
+
+    /**
+     * @return map of concept_fullname -> series_meta map, or null if not enabled or metadata not applicable
+     */
+    private Map<String,Map> computeMap(List<String> conceptPaths) {
+
+        //get all the OntologyTerms for the concepts
+        Set<OntologyTerm> terms = conceptPaths.collect {
+            conceptsResource.getByKey(getConceptKey(it))} as Set
+
+        //get all the SeriesMeta mapped by concept name
+        Map<String, Map> nameToSeriesMeta = terms.collectEntries {[it.fullName, it.metadata?.seriesMeta as Map]}
+
+        if (nameToSeriesMeta.size() > 0) {
+            String firstUnit = nameToSeriesMeta.values().first()?.unit?.toString()
+
+            //if all the units are the same and not null, and with numerical values
+            if (firstUnit != null &&
+                    nameToSeriesMeta.values().every { it?.value?.isInteger() && firstUnit == it?.unit?.toString() }) {
+
+                return nameToSeriesMeta
+            }
+        }
+
+        return null //nothing to return
+    }
+
+    public static String getConceptKey(String path) {
+        OpenHighDimensionalDataStep.createConceptKeyFrom(path)
+    }
+
 
 }
